requirements.txt
streamlit
tensorflow
plotly
pandas
numpy
Pillow
requests
geopy
googlemaps